backend/requirements.txt
# requirements.txt

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
starlette==0.41.2

# Database
sqlalchemy==2.0.36

# Security & Environment
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.2.1

# Google APIs
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0

# Scheduling
apscheduler==3.10.4

# HTTP & Networking
httpx==0.28.1
requests==2.32.3

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Utilities
aiofiles==24.1.0
python-dateutil==2.9.0

# Additional dependencies
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
h11==0.14.0
httpcore==1.0.7
idna==3.10
sniffio==1.3.1
typing-extensions==4.12.2
urllib3==2.3.0